--- services/director-v2/src/simcore_service_director_v2/components/monitoring.py
@@ -1,1 +1,0 @@
-# services/director/src/simcore_service_director/monitoring.py
