--- services/director-v2/src/simcore_service_director_v2/components/tracing.py
@@ -1,1 +1,0 @@
-## packages/service-library/src/servicelib/tracing.py
