+++ services/director-v2/src/simcore_service_director_v2/modules/monitoring.py
@@ -1,0 +1,7 @@
+# services/director/src/simcore_service_director/monitoring.py
+
+
+# Module's setup logic ---------------------------------------------
+
+
+# Module's business logic ---------------------------------------------
