+++ services/director-v2/src/simcore_service_director_v2/modules/tracing.py
@@ -1,0 +1,17 @@
+## packages/service-library/src/servicelib/tracing.py
+
+from pydantic import BaseSettings, AnyHttpUrl
+from ..core.settings import CommonConfig
+
+
+# Module's setup logic ---------------------------------------------
+class TracingSettings(BaseSettings):
+    enabled: bool = True
+    zipkin_endpoint: AnyHttpUrl = "http://jaeger:9411"
+
+    class Config(CommonConfig):
+        env_prefix = "TRACING_"
+
+
+
+# Module's business logic ---------------------------------------------
