--- services/director-v2/src/simcore_service_director_v2/services/director_v0.py
@@ -1,55 +1,0 @@
-import logging
-from dataclasses import dataclass
-from typing import Iterator
-
-import httpx
-from fastapi import FastAPI
-
-from ..core.settings import DirectorV0Settings
-from .client_utils import handle_response, handle_retry
-
-logger = logging.getLogger(__name__)
-
-
-def on_start(app: FastAPI) -> None:
-    settings: DirectorV0Settings = app.state.settings.director_v0
-    app.state.director_v0_client = DirectorV0Client(
-        client=httpx.AsyncClient(base_url=settings.base_url(include_tag=True))
-    )
-
-
-async def on_stop(app: FastAPI) -> None:
-    await app.state.director_v0_client.client.aclose()
-    del app.state.director_v0_client
-
-
-# TODO: add this functionlity
-async def on_cleanup_context(app: FastAPI) -> Iterator[None]:
-    settings: DirectorV0Settings = app.state.settings.director_v0
-    async with httpx.AsyncClient(
-        base_url=settings.base_url(include_tag=True)
-    ) as client:
-        app.state.director_v0_client = DirectorV0Client(client)
-
-        yield
-
-    del app.state.director_v0_client
-
-
-@dataclass
-class DirectorV0Client:
-    client: httpx.AsyncClient
-
-    @classmethod
-    def create(cls, app: FastAPI, **kwargs):
-        app.state.director_v0_client = cls(**kwargs)
-        return cls.instance(app)
-
-    @classmethod
-    def instance(cls, app: FastAPI):
-        return app.state.director_v0_client
-
-    @handle_response("Director", logger)
-    @handle_retry(logger)
-    async def request(self, method: str, tail_path: str, **kwargs):
-        return await self.client.request(method, tail_path, **kwargs)
